Removed zero ICs for Bid and Bad from module declarations
The problem was that though having a "placeholder" initial condition for, say
tBid made it convenient to do the kinds of things that you'd want to do in the
MOMP-only case, they made it very inconvenient when you wanted to embed the
module inside a larger model. In that situation, you would have to not only
override the initial value for the Parameter (e.g. Bid_0), you would also
have to reset the initial condition pattern to be associated with full-length,
rather than truncated, Bid. Since the initial conditions that were removed
were assigned zero values, people would have to override them with some
value anyway, so it seems reasonable to force people to have to add a condition
in.

The trade-off was that the initial conditions have to be declared for those
species to show up in the reaction network, and hence to produce the ODEs that
match the test conditions. To get around this, I had to modify the setup code
for every test to add Bid_0 and/or Bad_0 initial conditions as appropriate.