More fixes to pore_transport for shen models
Because the Cui models all use a dimeric "pore" rather than a tetramer, this required
that the shen_pore_transport macro be re-parameterized to take the pore_size as an
argument, which in turn required that the earm2.macros.pore_transport macro be
reparameterized to take the size as an argument (it had defaulted to 4).

To make the dimeric pore compatible with the dimeric species generated by pore_species,
I had to change the dimeric pore from

Bax(bf=None, s1=1, s2=2, state='A') % Bax(bf=None, s1=2, s2=1, state='A')

to

Bax(bf=None, s1=1, s2=None, state='A') % Bax(bf=None, s1=None, s2=1, state='A')

Interestingly, this changed the coefficient for the corresponding binding reaction
in the odes from being scaled by 1/2 to being scaled by 1. It turned out that this
is because, even though the binding reaction is still homomeric, because it can
happen in two ways (the Baxes can dimerize on either "side") it gets rescaled by 2,
and the two scaling factors cancel out (see comment in cui2008_direct in
shen_modules.py, below).

After making this fix, I had to edit the tested ODEs in test_shen_models.py
accordingly.