Removed Bad and Noxa (temporarily); made MATLAB for Hoda
Since setting the initial conditions to 0 doesn't remove Bad and Noxa from
the reaction network, the initial conditions had to be removed (since
Hoda should not be fitting them in this case). However even this left
the Bad and Noxa binding parameters in the parameter list, so they would
be included in the fitting procedure. So I had to comment out the sensitizer
binding step of the model. However, I plan to reinstate these elements
immediately after this commit.