Removed BclXL recruitment reactions
Previously, BidM and BaxA would bind BclXL-C, releasing it as BclXL-M. This
violated the semantics of C and M, as it ostensibly allowed mitochondrial
species (BidM and BaxA) to bind a cytosolic species (BclXL-C).

Allowing BclXL-C to have a translocation equilibrium to BclXL-M and then simply
having BaxA and BidM bind to BclXL-M would naturally give rise to the
phenomenon of BclXL translocating preferentially to the mitochondria upon
an apoptotic stimulus, without having to explicitly add in reactions for this
to happen. Hence I removed them from the model.