Some hacks to get shen_modules to pass with new units
The problem was that the test to verify the parameters meant that the
newly converted parameters would have to be converted back into the original
units to be verified. The problem was that the conversion process requires
knowing whether a parameter is a concentration, bimolecular rate constant,
unimolecular rate constant, etc. Basically it needs to know the units,
and right now we have no way of determining that except by the suffix on
the end of the parameter (e.g., '_kf'). This was problematic for displacement
reactions, which were previously listed as '_kr', even though the reverse
reaction is actually bimolecular in this case. It's also problematic because
the equilibrate macro produces a '_kf' and a '_kr' parameter, even though
the _kf is in this case actually unimolecular! I added a special case hack for
this for the time being.

I changed the names produced by some of the macros, but this also meant that
I had to update the parameter name lists for all of the models.

Ultimately this will have to be fixed to keep track of the real units for
parameters.